[RL] moved ResourceModels to new directory + updated paths
