+++ ResourceLib/BelowGround/Generic/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Nov  8 15:25:03 2018
+
+@author: bathmann
+"""
+
+from .OGSLargeScale3DExternal import OGSLargeScale3DExternal
+from .NetworkOGSLargeScale3DExternal import NetworkOGSLargeScale3DExternal
