+++ ResourceLib/BelowGround/Network/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Nov  8 15:25:03 2018
+
+@author: bathmann
+"""
+
+from .SimpleNetwork import SimpleNetwork
+from .NetworkFixedSalinity import NetworkFixedSalinity
+from .NetworkOGSLargeScale3D import NetworkOGSLargeScale3D
