+++ ResourceLib/ResourceModel.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+@date: 2018-Today
+@author: jasper.bathmann@ufz.de
+"""
+
+
+class ResourceModel:
+
+    def getAbovegroundResources(self):
+        return self.aboveground_resources
+
+    def getBelowgroundResources(self):
+        return self.belowground_resources
