--- TreeModelLib/BelowgroundCompetition/__init__.py
@@ -1,21 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Nov  8 15:25:03 2018
-
-@author: bathmann
-"""
-
-from .SimpleTest import SimpleTest
-from .FON import FON
-from .OGSWithoutFeedback import OGSWithoutFeedback
-from .OGSLargeScale3D import OGSLargeScale3D
-from .OGS.helpers import CellInformation
-from .FixedSalinity import FixedSalinity
-from .SimpleNetwork import SimpleNetwork
-from .NetworkFixedSalinity import NetworkFixedSalinity
-from .NetworkOGSLargeScale3D import NetworkOGSLargeScale3D
-from .OGSLargeScale3DExternal import OGSLargeScale3DExternal
-from .NetworkOGSLargeScale3DExternal import NetworkOGSLargeScale3DExternal
-from .SymmetricZOI import SymmetricZOI
-from .SZoiFixedSalinity import SZoiFixedSalinity
